cephalometric_mvp/backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0  # For SQLite async support (demo mode)
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ML and Image Processing
torch>=2.1.0
torchvision>=0.16.0
segmentation-models-pytorch>=0.3.3
Pillow>=10.2.0
numpy>=1.26.0
opencv-python-headless>=4.9.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
